requirements.txt
python-twitter
numpy
pyshorteners
python-dotenv
pyquery
requests

# polyglot
# pycld2
# morfessor
# bitly_api
